ExportAction: -fixed possible npe when nothing is selected ExtractAction: -updated FileFilter to only show first part of archives UnixSplit: -updated isSupported, did accept every file extension
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@45132 ebf7c1c2-ba36-0410-9fe8-c592906822b4

Former-commit-id: df631343c2b8b81b800c4d3f316ea4e6d60ff657

+++ src/jd/gui/swing/jdgui/views/settings/panels/linkgrabberfilter/ExportAction.java
@@ -69,17 +69,19 @@ public class ExportAction extends AppAction {
             d.setMultiSelection(false);
             Dialog.I().showDialog(d);
             File saveto = d.getSelectedFile();
-            if (!saveto.getName().endsWith(ext)) {
-                saveto = new File(saveto.getAbsolutePath() + ext);
-            }
-            try {
-                if (saveto.exists() && !saveto.delete()) {
-                    throw new IOException("Could not delete/overwrite:" + saveto);
+            if (saveto != null) {
+                if (!saveto.getName().endsWith(ext)) {
+                    saveto = new File(saveto.getAbsolutePath() + ext);
                 }
-                IO.writeStringToFile(saveto, JSonStorage.serializeToJson(export));
-            } catch (IOException e1) {
-                LogController.CL().log(e1);
-                Dialog.getInstance().showExceptionDialog(_GUI.T.lit_error_occured(), e1.getMessage(), e1);
+                try {
+                    if (saveto.exists() && !saveto.delete()) {
+                        throw new IOException("Could not delete/overwrite:" + saveto);
+                    }
+                    IO.writeStringToFile(saveto, JSonStorage.serializeToJson(export));
+                } catch (IOException e1) {
+                    LogController.CL().log(e1);
+                    Dialog.getInstance().showExceptionDialog(_GUI.T.lit_error_occured(), e1.getMessage(), e1);
+                }
             }
         } catch (DialogNoAnswerException e1) {
             LogController.CL().log(e1);
